change addon to embedded
