+++ demo/yun/pages/moments/index.md
@@ -1,0 +1,9 @@
+---
+layout: moments
+title: 小随想
+description: 记录生活里的小事
+toc: false
+moments:
+  initialCount: 10
+  batchSize: 10
+---
