+++ demo/yun/valaxy.config.ts
@@ -7,7 +7,6 @@ import { addonBangumi } from 'valaxy-addon-bangumi'
 import { addonComponents } from 'valaxy-addon-components'
 import { addonGirls } from 'valaxy-addon-girls'
 import { addonLightGallery } from 'valaxy-addon-lightgallery'
-import { addonMoments } from 'valaxy-addon-moments'
 import { addonTest } from 'valaxy-addon-test'
 
 import { localIconLoader } from 'vitepress-plugin-group-icons'
@@ -118,12 +117,6 @@ export default defineValaxyConfig<ThemeConfig>({
     //   },
     // }),
     addonTest(),
-    addonMoments({
-      title: '小随想',
-      description: '记录生活里的小事',
-      initialCount: 10,
-      batchSize: 10,
-    }),
   ],
 
   groupIcons: {
