--- packages/valaxy-addon-moments/client/options.ts
@@ -1,7 +1,0 @@
-import type { MomentsOptions } from '../types'
-import { useAddonConfig } from 'valaxy'
-import pkg from '../package.json'
-
-export function useMomentsConfig() {
-  return useAddonConfig<MomentsOptions>(pkg.name)
-}
