--- packages/valaxy-addon-moments/index.d.ts
@@ -1,3 +1,0 @@
-export * from './client'
-export * from './node'
-export * from './types'
