--- packages/valaxy-addon-moments/locales/en.yml
@@ -1,14 +1,0 @@
-addon:
-  moments:
-    author: Author
-    collapse: Collapse
-    empty: No moments yet.
-    expand: Expand
-    image: Moment image {index}
-    images: '{count} attached images'
-    like: Like this moment
-    more: Load more ({count})
-    month: month
-    pinned: Pinned
-    timeline: Timeline
-    unlike: Unlike this moment
