--- packages/valaxy-addon-moments/locales/zh-CN.yml
@@ -1,14 +1,0 @@
-addon:
-  moments:
-    author: 作者
-    collapse: 收起
-    empty: 还没有发布动态。
-    expand: 展开
-    image: 动态图片 {index}
-    images: 共 {count} 张图片
-    like: 点赞
-    more: 加载更多（剩余 {count} 条）
-    month: 月
-    pinned: 置顶
-    timeline: 时间线
-    unlike: 取消点赞
