--- packages/valaxy-addon-moments/pages/moments/index.vue
@@ -1,10 +1,0 @@
-<template>
-  <ValaxyMoments>
-    <template v-if="$slots.header" #header="headerProps">
-      <slot
-        name="header"
-        v-bind="headerProps"
-      />
-    </template>
-  </ValaxyMoments>
-</template>
