+++ packages/valaxy-theme-yun/layouts/moments.vue
@@ -8,27 +8,29 @@ const yun = useYunAppStore()
   <YunLayoutWrapper>
     <YunLayoutLeft />
 
-    <main class="yun-main yun-moments-main lt-md:w-full" flex="~ center">
-      <div class="content w-full md:w-3xl lg:w-2xl xl:w-2xl 2xl:w-4xl">
-        <RouterView v-slot="{ Component }">
-          <component :is="Component">
-            <template #header="{ description, title }">
-              <div class="yun-moments-header">
-                <YunPageHeader :title="title" />
-                <p v-if="description">
-                  {{ description }}
-                </p>
-              </div>
-            </template>
-          </component>
-        </RouterView>
-      </div>
-    </main>
+    <RouterView v-slot="{ Component }">
+      <component :is="Component">
+        <template #main>
+          <div class="content yun-moments-main w-full md:w-3xl lg:w-2xl xl:w-2xl 2xl:w-4xl">
+            <ValaxyMoments>
+              <template #header="{ description, title }">
+                <div class="yun-moments-header">
+                  <YunPageHeader :title="title" />
+                  <p v-if="description">
+                    {{ description }}
+                  </p>
+                </div>
+              </template>
+            </ValaxyMoments>
+          </div>
+        </template>
+      </component>
+    </RouterView>
 
     <button
       class="xl:hidden toc-btn shadow-md fixed yun-icon-btn z-20 bg-$va-c-bg-soft"
       opacity="75" right="4" bottom="19"
-      :aria-label="$t('addon.moments.timeline')"
+      :aria-label="$t('moments.timeline')"
       @click="yun.rightSidebar.toggle()"
     >
       <div i-ri-time-line />
@@ -64,11 +66,5 @@ const yun = useYunAppStore()
   margin: -0.25rem 0 0;
   color: var(--va-c-text-2);
   text-align: center;
-}
-
-@media (width < 768px) {
-  .yun-moments-header {
-    padding-inline: 1rem;
-  }
 }
 </style>
