+++ packages/valaxy/client/components/ValaxyMomentsTimeline.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
-import { groupMomentsByYear, useMoments } from '../client'
+import { groupMomentsByYear, useMoments } from '../composables/moments'
 
 const { t } = useI18n()
 const moments = useMoments()
@@ -16,10 +16,10 @@ function navigate(anchor: string) {
 
 <template>
   <div class="valaxy-moments-timeline">
-    <slot name="title" :title="t('addon.moments.timeline')">
-      <h2>{{ t('addon.moments.timeline') }}</h2>
+    <slot name="title" :title="t('moments.timeline')">
+      <h2>{{ t('moments.timeline') }}</h2>
     </slot>
-    <nav :aria-label="t('addon.moments.timeline')">
+    <nav :aria-label="t('moments.timeline')">
       <section v-for="year in years" :key="year.year">
         <strong>{{ year.year }}</strong>
         <button
@@ -28,7 +28,7 @@ function navigate(anchor: string) {
           type="button"
           @click="navigate(month.anchor)"
         >
-          <span>{{ month.label }} {{ t('addon.moments.month') }}</span>
+          <span>{{ month.label }} {{ t('moments.month') }}</span>
           <small>{{ month.moments.length }}</small>
         </button>
       </section>
