+++ packages/valaxy/client/composables/moments/index.ts
@@ -1,5 +1,4 @@
 export * from './data'
-export * from './options'
 export * from './time'
 export * from './useMomentLike'
 export * from './useMoments'
