+++ packages/valaxy/client/layouts/moments.vue
@@ -1,0 +1,9 @@
+<template>
+  <RouterView v-slot="{ Component }">
+    <component :is="Component">
+      <template #main-content>
+        <ValaxyMoments />
+      </template>
+    </component>
+  </RouterView>
+</template>
