+++ packages/valaxy/node/plugins/index.ts
@@ -1,3 +1,2 @@
-export { createLightMarkdownRenderer } from './markdown'
 export * from './preset'
 export * from './valaxy'
