+++ packages/valaxy/types/index.ts
@@ -11,5 +11,6 @@ export * from './frontmatter'
 
 // used in node, but without node deps
 // and need be shared with client types
+export * from './moments'
 export * from './node'
 export * from './posts'
